CI: Add conventional commit checker
ci: add new ci to current ci jobs

fix: linting error related to new CI

ci: clean the code

chore: add contributor

ci: fix linting errors

ci: fix error when PR from fork

ci: fix linting errors
