--- .github/workflows/lint_commits.yml
@@ -1,26 +1,0 @@
-name: Lint Commits
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-
-jobs:
-  lint_commits:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: '3.x'
-
-      - name: Validate commit messages
-        run: |
-          git log --pretty=format:"%s" | head -n 1 | python .github/scripts/validate-commit-msg.py
