+++ AUTHORS.md
@@ -10,3 +10,4 @@
 - Rami, GOUAL, etudes.ramigoual@gmail.fr
 - Samy FERGUI, samy.fergui.pro@gmail.com
 - Marwane, RACHAD, marwanerachad@gmail.com
+- Mathis, Beauville, mathis.beauville@etud.univ-evry.fr
